Split large server file into multiple to improve readability.
